+++ volatility/models/GarmanKlass.py
@@ -2,7 +2,7 @@ import math
 
 import numpy as np
 
-from volatility.models.api import get_variance_overlapping_adjustment_factor
+from volatility.models.overlapping_sample import get_variance_overlapping_adjustment_factor
 
 
 def get_estimator(price_data, window=30, trading_periods=252, clean=True, use_overlapping_adjustment_factor=True):
@@ -12,7 +12,7 @@ def get_estimator(price_data, window=30, trading_periods=252, clean=True, use_ov
     log_hl = (price_data['High'] / price_data['Low']).apply(np.log)
     log_co = (price_data['Close'] / price_data['Open']).apply(np.log)
 
-    rs = 0.5 * log_hl**2 - (2*math.log(2)-1) * log_co**2
+    rs = 0.5 * (log_hl**2) - (2*math.log(2)-1) * (log_co**2)
     
     def f(v):
         return ((trading_periods * v.mean())**0.5) * math.sqrt(m)
